requirements.txt
pillow
praw
python-dotenv
pyrae
qrcode
setproctitle
cowpy
pyyaml
hikari[speedups]
hikari-lightbulb
uvloop
pytest
graphviz
tweepy
colorama
apscheduler
coloredlogs
sqlite